refactor(lens): migrate get_discovery_summary tool to free function (epic #1655 PR 1/9)
First slice of the Executor tool migration. Moves _tool_get_discovery_summary
from executor.py to a top-level get_discovery_summary() in the domain's
registration file, inlines the previously-nested _risk_level helper at
module level, and rewires the ToolDef impl directly (no more _impl shim
indirection or per-call SessionLocal from the shim).

Behavior-preserving: live smoke against local Postgres confirms identical
return shape ({total, under_guard, missing, coverage_pct, by_framework,
high_risk, agents}).

Proves the pattern including the nested-helper case. Next: policies (2 tools).
